+++ src/content/src/Tests/WebAPIService.Unit.Tests/UnitTest1.cs
@@ -1,6 +1,3 @@
-using System.Net;
-using System.Net.Http;
-using System.Threading.Tasks;
 using Eshopworld.Tests.Core;
 using Xunit;
 
@@ -8,12 +5,8 @@ using Xunit;
 public class UnitTest1
 {
     [Fact, IsUnit]
-    public async Task Test1()
+    public void Test1()
     {
-        var httpClient = new HttpClient();
-        var response = await httpClient.GetAsync("https://security-sts.production.eshopworld.com/.well-known/openid-configuration");
-
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
 
     }
 }
